feat: api handle redirects (#3246)
* feat: api handle redirects

* broken code

* comment out tests

* tests

* tests

* remove

* directory for username

* directory for links

* api redirect

* changes

* Restore package-lock.json
